requirements.txt
Django==2.2.16
mixer==7.1.2
Pillow==8.3.1
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
requests==2.26.0
six==1.16.0
sorl-thumbnail==12.7.0
Faker==12.0.1

py~=1.11.0
colorama~=0.4.4
certifi~=2022.5.18.1
pip~=21.1.2
wheel~=0.36.2
pytz~=2022.1
toml~=0.10.2
attrs~=21.4.0
python-dateutil~=2.8.2
sqlparse~=0.4.2
wcwidth~=0.2.5
pluggy~=0.13.1
iniconfig~=1.1.1
atomicwrites~=1.4.0
idna~=2.8
urllib3~=1.25.11
chardet~=3.0.4
pyparsing~=3.0.9
setuptools~=57.0.0
django-debug-toolbar~=3.2.4